feat(adapi_specs): rename control command api
